[Foundation-Fixit] Use python3 in PRESUBMIT.py.
This also switches remaining pylint-1.5 presubmit (for 'components')
to use (default) pylint-2.7 and updates all corresponding lint
directives.

Some non-critical very noisy or confusing lints are disabled
right in PRESUBMIT.py instead of individual files. This is
arguably easier to maintain (especially when old lints used
cryptic names like W0401).

R=iannucci@chromium.org

Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/3984375
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>

+++ appengine/components/components/auth/ipaddr.py
@@ -74,7 +74,7 @@ def ip_from_string(ipstr):
       raise ValueError('not IPv4 or IPv6 address')
   except ValueError as e:
     raise ValueError('%r is not an IP address (%s)' % (ipstr, e))
-  value = 0L
+  value = long(0)
   for i in values:
     value = value * factor + i
   return IP(bits, value)
